+++ src/dialects/mssql/index.js
@@ -43,7 +43,7 @@ assign(Client_MSSQL.prototype, {
   },
 
   formatter() {
-    return new MSSQL_Formatter(this)
+    return new MSSQL_Formatter(this, ...arguments)
   },
 
   transaction() {
@@ -67,7 +67,7 @@ assign(Client_MSSQL.prototype, {
   },
 
   wrapIdentifierImpl(value) {
-    return (value !== '*' ? `[${value.replace(/\[/g, '\[')}]` : '*')
+    return (value !== '*' ? `[${value.replace(/\[/g, '[')}]` : '*')
   },
 
   // Get a raw connection, called by the `pool` whenever a new