+++ src/query/string.js
@@ -1,6 +1,6 @@
 /*eslint max-len: 0, no-var:0 */
 
-export const charsRegex = /[\0\b\t\n\r\x1a\"\'\\]/g; // eslint-disable-line no-control-regex
+export const charsRegex = /[\0\b\t\n\r\x1a"'\\]/g; // eslint-disable-line no-control-regex
 export const charsMap = {
   '\0': '\\0',
   '\b': '\\b',
@@ -153,7 +153,7 @@ function convertTimezone(tz) {
   if (tz === 'Z') {
     return 0;
   }
-  const m = tz.match(/([\+\-\s])(\d\d):?(\d\d)?/);
+  const m = tz.match(/([+\-\s])(\d\d):?(\d\d)?/);
   if (m) {
     return (m[1] == '-' ? -1 : 1) * (parseInt(m[2], 10) + ((m[3] ? parseInt(m[3], 10) : 0) / 60)) * 60;
   }