+++ test/integration/builder/additional.js
@@ -13,8 +13,9 @@ module.exports = function(knex) {
     describe("Custom response processing", () => {
 
       before('setup custom response handler', () => {
-        knex.client.config.postProcessResponse = (response) => {
+        knex.client.config.postProcessResponse = (response, queryContext) => {
           response.callCount = response.callCount ? (response.callCount + 1) : 1;
+          response.queryContext = queryContext;
           return response;
         };
       });
@@ -29,9 +30,27 @@ module.exports = function(knex) {
         });
       });
 
+      it('should pass query context to the custom handler', () => {
+        return knex('accounts')
+          .queryContext('the context')
+          .limit(1)
+          .then(res => {
+            expect(res.queryContext).to.equal('the context');
+          });
+      });
+
       it('should process raw response', () => {
         return knex.raw('select * from ??', ['accounts']).then(res => {
+          expect(res.callCount).to.equal(1);
         });
+      });
+
+      it('should pass query context for raw responses', () => {
+        return knex.raw('select * from ??', ['accounts'])
+          .queryContext('the context')
+          .then(res => {
+            expect(res.queryContext).to.equal('the context');
+          });
       });
 
       it('should process response done in transaction', () => {
@@ -44,6 +63,53 @@ module.exports = function(knex) {
           expect(res.callCount).to.equal(1);
         });
       });
+
+      it('should pass query context for responses from transactions', () => {
+        return knex.transaction(trx => {
+          return trx('accounts')
+            .queryContext('the context')
+            .limit(1)
+            .then(res => {
+              expect(res.queryContext).to.equal('the context');
+              return res;
+            });
+        }).then(res => {
+          expect(res.queryContext).to.equal('the context');
+        });
+      });
+    });
+
+    describe('columnInfo with wrapIdentifier and postProcessResponse', () => {
+
+      before('setup hooks', () => {
+        knex.client.config.postProcessResponse = (response) => {
+          return _.mapKeys(response, (val, key) => {
+            return _.camelCase(key);
+          });
+        };
+
+        knex.client.config.wrapIdentifier = (id, origImpl) => {
+          return origImpl(_.snakeCase(id));
+        };
+      });
+
+      after('restore client configuration', () => {
+        knex.client.config.postProcessResponse = null;
+        knex.client.config.wrapIdentifier = null;
+      });
+
+      it('should work using camelCased table name', () => {
+        return knex('testTableTwo').columnInfo().then(res => {
+          expect(Object.keys(res)).to.eql(['id', 'accountId', 'details', 'status', 'jsonData']);
+        });
+      });
+
+      it('should work using snake_cased table name', () => {
+        return knex('test_table_two').columnInfo().then(res => {
+          expect(Object.keys(res)).to.eql(['id', 'accountId', 'details', 'status', 'jsonData']);
+        });
+      });
+
     });
 
     it('should forward the .get() function from bluebird', function() {
@@ -98,16 +164,30 @@ module.exports = function(knex) {
           tester('oracle', "truncate table \"test_table_two\"");
           tester('mssql', 'truncate table [test_table_two]');
         })
-        .then(function() {
-
+        .then(() => {
           return knex('test_table_two')
             .select('*')
-            .then(function(resp) {
+            .then(resp => {
               expect(resp).to.have.length(0);
             });
-
+        })
+        .then(() => {
+          // Insert new data after truncate and make sure ids restart at 1.
+          // This doesn't currently work on oracle, where the created sequence
+          // needs to be manually reset.
+          if (knex.client.dialect !== 'oracle') {
+            return knex('test_table_two').insert({ status: 1 })
+              .then(res => {
+                return knex('test_table_two')
+                  .select('id')
+                  .first()
+                  .then(res => {
+                    expect(res).to.be.an('object')
+                    expect(res.id).to.equal(1);
+                  });
+              });
+          }
         });
-
     });
 
     it('should allow raw queries directly with `knex.raw`', function() {