+++ test/integration/builder/transaction.js
@@ -5,6 +5,7 @@
 var Promise = testPromise;
 var Knex   = require('../../../knex');
 var _ = require('lodash');
+var sinon = require('sinon')
 
 module.exports = function(knex) {
 
@@ -174,18 +175,18 @@ module.exports = function(knex) {
       if (knex.client.dialect === 'postgresql') {
         return knex.transaction(function(trx) {
           return trx.schema.createTable('test_schema_transactions', function(table) {
-              table.increments();
-              table.string('name');
-              table.timestamps();
-            }).then(function() {
-              return trx('test_schema_transactions').insert({name: 'bob'});
-            }).then(function() {
-              return trx('test_schema_transactions').count('*');
-            }).then(function(resp) {
-              var _count = parseInt(resp[0].count, 10);
-              expect(_count).to.equal(1);
-              throw err;
-            });
+            table.increments();
+            table.string('name');
+            table.timestamps();
+          }).then(function() {
+            return trx('test_schema_transactions').insert({name: 'bob'});
+          }).then(function() {
+            return trx('test_schema_transactions').count('*');
+          }).then(function(resp) {
+            var _count = parseInt(resp[0].count, 10);
+            expect(_count).to.equal(1);
+            throw err;
+          });
         })
         .on('query', function(obj) {
           count++;
@@ -288,7 +289,6 @@ module.exports = function(knex) {
       })
       .then(expectQueryEventToHaveBeenTriggered)
       .catch(expectQueryEventToHaveBeenTriggered);
-
     });
 
     it('#1040, #1171 - When pool is filled with transaction connections, Non-transaction queries should not hang the application, but instead throw a timeout error', function() {
@@ -437,7 +437,35 @@ module.exports = function(knex) {
         .spread(function (ret1, ret2) {
           expect(ret1).to.equal(ret2);
         });
-     });
+    });
 
+    it('should pass the query context to wrapIdentifier', function () {
+      const originalWrapIdentifier = knex.client.config.wrapIdentifier;
+      const spy = sinon.spy().named('calledWithContext');
+
+      function restoreWrapIdentifier() {
+        knex.client.config.wrapIdentifier = originalWrapIdentifier;
+      }
+
+      knex.client.config.wrapIdentifier = (value, wrap, queryContext) => {
+        spy(queryContext);
+        return wrap(value);
+      };
+
+      return knex.transaction(function(trx) {
+        return trx
+          .select()
+          .from('accounts')
+          .queryContext({ foo: 'bar' });
+      }).then(function() {
+        expect(spy.callCount).to.equal(1);
+        expect(spy.calledWith({ foo: 'bar' })).to.equal(true);
+      }).then(function() {
+        restoreWrapIdentifier();
+      }).catch(function(e) {
+        restoreWrapIdentifier();
+        throw e;
+      });
+    });
   });
 };